requirements.txt
django
django-crispy-forms
django-environ
djangorestframework
kafka-python
psycopg2-binary==2.8.5
pytest
pytest-mock
python3-wget
kafka-python
requests
scrapy
validators
